backend/requirements.txt
fastapi==0.115.0
uvicorn==0.32.0
python-multipart==0.0.12
httpx==0.27.2
nbformat==5.10.4
python-dotenv==1.0.1
google-adk>=0.5.0
google-genai>=0.8.0
